Machine-Learning-App/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
blinker==1.6.2
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
confection==0.1.0
cymem==2.0.7
Flask==2.3.2
Flask-Login==0.6.2
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
gunicorn==21.2.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
keras==2.13.1
langcodes==3.3.0
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
pathy==0.10.2
preshed==3.0.8
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
smart-open==6.3.0
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
#tensorflow-macos==2.13.0
termcolor==2.3.0
textblob==0.17.1
textstat==0.7.3
thinc==8.1.10
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
